datapath: move RemoteSNATDstAddr* as methods on pkg/node.LocalNode
